~~~ - Server.Packets changed to ServerPackts (static) - Body animations are properly working - Utility.Clamp very useful - Extensions.ApplySpell / ApplySkill also very useful - More things stored inside the panelobjects - Slight changes to creation engine objects, methods, and delegates - Whug is now a Guinea Pig - UserOptions is fixed - Activations are somewhat more universalized - Changes starting map, lr6 verypretty         *Will need to create a large enum for animations - Replaced TryGetObject methods in world, with far superior versions - Properly locked the rest of the objects - Converted some dictionaries to Lists of KeyValuPairs (for order) - Added loopback support (because my router doesnt have nat loopback fml) - Added more client shorthand - PanelObjects have effectanimation and bodyanimation - Fixed EquipmentSlot swap in profile user equipment display - Disabled some expensive code in warpUser       *(need to implement some sort of radial search with a yield return)
----EXCHANGE GOLD IS BROKEN, SHOULD FIX ASAP

+++ Chaos/Objects/Gold.cs
@@ -15,7 +15,7 @@ namespace Chaos
     {
         internal uint Amount { get; set; }
 
-        internal Gold(byte sprite, Point point, Map map, uint amount)
+        internal Gold(ushort sprite, Point point, Map map, uint amount)
           : base((ushort)(sprite + CONSTANTS.ITEM_SPRITE_OFFSET), point, map)
         {
             Amount = amount;
